Fix deleteUserById

+++ graphql/resolvers/user.js
@@ -220,10 +220,11 @@ module.exports = {
             const currentUser = await User.findById(currentId);
             if (currentUser.role) {
                 const user = await User.findById(userId);
-                await user.delete();
-                return true;
-            }
-            throw new ApolloError('权限不足！');
+                if (user) {
+                    await user.delete();
+                    return true;
+                } else throw new ApolloError('用户不存在！');
+            } else throw new ApolloError('权限不足！');
         }
     }
 };
